+++ entities/character.go
@@ -1,0 +1,7 @@
+package entities
+
+type Character struct {
+	Name  string
+	Race  *Race
+	Class *Class
+}
